Extract mempool common functionality to BaseMempool
