+++ bchain/mempool_bitcoin_type.go
@@ -1,37 +1,16 @@
 package bchain
 
 import (
-	"sort"
-	"sync"
 	"time"
 
 	"github.com/golang/glog"
 )
 
-type addrIndex struct {
-	addrDesc string
-	n        int32
-}
-
-type txEntry struct {
-	addrIndexes []addrIndex
-	time        uint32
-}
-
-type txidio struct {
-	txid string
-	io   []addrIndex
-}
-
 // MempoolBitcoinType is mempool handle.
 type MempoolBitcoinType struct {
-	chain               BlockChain
-	mux                 sync.Mutex
-	txEntries           map[string]txEntry
-	addrDescToTx        map[string][]Outpoint
+	BaseMempool
 	chanTxid            chan string
 	chanAddrIndex       chan txidio
-	OnNewTxAddr         OnNewTxAddrFunc
 	AddrDescForOutpoint AddrDescForOutpointFunc
 }
 
@@ -39,7 +18,9 @@ type MempoolBitcoinType struct {
 // For now there is no cleanup of sync routines, the expectation is that the mempool is created only once per process
 func NewMempoolBitcoinType(chain BlockChain, workers int, subworkers int) *MempoolBitcoinType {
 	m := &MempoolBitcoinType{
-		chain:         chain,
+		BaseMempool: BaseMempool{
+			chain: chain,
+		},
 		chanTxid:      make(chan string, 1),
 		chanAddrIndex: make(chan txidio, 1),
 	}
@@ -66,30 +47,6 @@ func NewMempoolBitcoinType(chain BlockChain, workers int, subworkers int) *Mempo
 	}
 	glog.Info("mempool: starting with ", workers, "*", subworkers, " sync workers")
 	return m
-}
-
-// GetTransactions returns slice of mempool transactions for given address
-func (m *MempoolBitcoinType) GetTransactions(address string) ([]Outpoint, error) {
-	parser := m.chain.GetChainParser()
-	addrDesc, err := parser.GetAddrDescFromAddress(address)
-	if err != nil {
-		return nil, err
-	}
-	return m.GetAddrDescTransactions(addrDesc)
-}
-
-// GetAddrDescTransactions returns slice of mempool transactions for given address descriptor
-func (m *MempoolBitcoinType) GetAddrDescTransactions(addrDesc AddressDescriptor) ([]Outpoint, error) {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	return append([]Outpoint(nil), m.addrDescToTx[string(addrDesc)]...), nil
-}
-
-func (m *MempoolBitcoinType) updateMappings(newTxEntries map[string]txEntry, newAddrDescToTx map[string][]Outpoint) {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	m.txEntries = newTxEntries
-	m.addrDescToTx = newAddrDescToTx
 }
 
 func (m *MempoolBitcoinType) getInputAddress(input Outpoint) *addrIndex {
@@ -221,36 +178,4 @@ func (m *MempoolBitcoinType) Resync() (int, error) {
 	m.updateMappings(newTxEntries, newAddrDescToTx)
 	glog.Info("mempool: resync finished in ", time.Since(start), ", ", len(m.txEntries), " transactions in mempool")
 	return len(m.txEntries), nil
-}
-
-func (a MempoolTxidEntries) Len() int      { return len(a) }
-func (a MempoolTxidEntries) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a MempoolTxidEntries) Less(i, j int) bool {
-	// if the Time is equal, sort by txid to make the order defined
-	hi := a[i].Time
-	hj := a[j].Time
-	if hi == hj {
-		return a[i].Txid > a[j].Txid
-	}
-	// order in reverse
-	return hi > hj
-}
-
-func getAllEntries(txEntries map[string]txEntry) MempoolTxidEntries {
-	a := make(MempoolTxidEntries, len(txEntries))
-	i := 0
-	for txid, entry := range txEntries {
-		a[i] = MempoolTxidEntry{
-			Txid: txid,
-			Time: entry.time,
-		}
-		i++
-	}
-	sort.Sort(a)
-	return a
-}
-
-// GetAllEntries returns all mempool entries sorted by fist seen time in descending order
-func (m *MempoolBitcoinType) GetAllEntries() MempoolTxidEntries {
-	return getAllEntries(m.txEntries)
 }
